// src/domain/enums.ts
export enum USER_GENDER {
    MALE = 'MALE',
    FEMALE = 'FEMALE',
    OTHER = 'OTHER',
}

export enum USER_ROLE {
    ADMIN = 'ADMIN',
    USER = 'USER',
    CLIENT_USER = 'CLIENT',
}

export enum SERVICE_CATEGORY{
    HAIRCUT = 'HAIRCUT',
    MAKEUP = 'MAKEUP',
    EYEBROWS = 'EYEBROWS',
    NAILS = 'NAILS',
    HAIRSTYLE = 'HAIRSTYLE',
    MASSAGE = 'MASSAGE',
    DEPILATION = 'DEPILATION',
    FACIAL_TREATMENT = 'FACIAL_TREATMENT',
    MULTISERVICES = 'MULTISERVICES'
}